[amber] rate limit should be a uint64

+++ go/src/amber/ipcserver/control_server.go
@@ -84,7 +84,7 @@ func (c *ControlSrvr) AddSrc(cfg amber.SourceConfig) (bool, error) {
 	}
 
 	tufSource := source.NewTUFSource(cfg.RepoUrl, dir, keys,
-		time.Millisecond*time.Duration(cfg.RatePeriod), uint64(cfg.RateLimit))
+		time.Millisecond*time.Duration(cfg.RatePeriod), cfg.RateLimit)
 	c.daemon.AddSource(tufSource)
 
 	return true, nil
